"Fix" Counter for LairBot? Other various adjustments and tweaks. - Don't forget to update usernames. - Subscribe to a ReactionAdded event instead of spamming message updates. - Fix $peek for WiFi? - Stop automatic builds for PRs. - Sanitize spaces in species names, lock file reading and writing. - Minor clean. Take 3 - Ignore charge moves when considering an upgrade for OHKO LairBot. - Allow custom file path for EncounterLogs. - Beautify EncounterLogs and log more info. - Fix oops with dex entry message when gifting. - Fix oops with form indexing. - Add option to clear buddy's nickname. - Expand DexFlavor to include forms and Gmax descriptions. - Add basic buddy leveling system on successful catches. - Add option to clear active buddy. - Override game version for some legends/UBs to UM/US if square shiny was rolled. - Add rudimentary buddy system with nicknaming and hatching. - Smallify Lair Embeds. - Remove extra showdown set line for language if Emerald Mew was rolled. - Check if form is valid in SwSh during typed event rng. - Fix incorrect form argument syntax during set generation. - Fix accidental Kanto bird shiny lock. - Tweak catching (apply Silvally and Genesect forms early). - Terminate early on failed set regen. - Handle "failed" catch better. - Minor tweaks to catch rng. - Remove perk command input check. - Add try catch for TradeCord command processing, use static UserInfo for serializing instead of passing it as an argument. - Minor compatibility changes for main's commits. - Rewrite and refactor TradeCord. - Disallow multiple TradeCord instances (serialization safety). - Add $peek <address> command (grabs screenshots from a specific console). - Change USB's read method to always account for large reads. - Minor clean. - Add community event vote. - Add command to delete user's data. - Use BlockingCollection for thread-safe command processing. - Prevent bots from receiving gifts. - Handle multiple addresses for $on and $off commands properly. - LairBot now requires a user to find their screen offsets. - Partial LairBot rewrite to accommodate dock. - Fix LairBot notes and caught flag reset. - Condense and trim offsets used by LairBot. - Fix other older LairBot bugs. - Add self bot scanner for TradeCord. - Slightly improve json safety? - Add console screen on/off Discord command. - Some TradeCord bug fixes. - Don't initialize large resources on startup if not necessary. - Fix oops with missing TradeCord and FixOT routines. - Only allow hex seed input for DenBot's seed injection. - Band-aid fix for bugged Keldeo, tweak IV RNG yet again. - Change DaySkip calls, fix AutoRoll overskipping for some, directly call ResetTime, fix Ditto and multi-turn moves for LairBot. - Use fixed read method for large data. - Remove unused methods and classes. - Add OHKO-less routine for LairBot. (#28)
- Add a rough weighted move selection logic.
- Improve the "upgrade" routine, make it an option.
- Add an option to keep a path in OHKO-less mode by resetting the game.
- Improve battle logic.
- Improve Dynamax routine.
- Account for some edge case battle interactions and status conditions.
- Change a lot of the offsets used.
- Take certain party shenanigans into account.
Enforce a sane config update interval.
Also test Json integrity before making a backup file.
Add missing Dutch language to DenBot.
Clean up TradeCord's randomizer.
Fix memes if item wasn't recognized for ItemTrade.
Fix unsafe variable assignment for DenBot.
Fix wild encounter IV randomizer.
Use bot's in-game name for TradeStart messages.
-Read distribution raids from memory instead of manually updating Json files.
-Band-aid fix for USB-Botbase's inability to read large data.
-When using AutoRoll, only read event data and tables when necessary.
Fix Cursola in daycare.
Remove one command at a time, add a sudo "$cc" command that removes a user's command entry (fallback in case a certain bug is still not fixed).
Use user ID from config instead of Context.
Allow gifting to add to Dex if Dex hasn't been completed, or if fewer than 50 entries left.
Stop showing default forms (like "Mimikyu-Disguised").
Fix some text output.
Add more Go encounters (Celebi, Jirachi, Genesect, Victini).
Fix Nidoran set generation issues, misc changes.
Fix massrelease without optional parameters.
Fix massrelease by ball type message.
Add ability to massrelease by ball type.
Add a perk system for TradeCord.
Use smaller ball images for TradeCord (resized and uploaded by BakaKaito).
Add option to set ball image as Author Icon url.
Remove wild space.
Fix Egg IDs not displayed with meme catches.
Add option to use either full-size catch images, or jan0809's downsized ones, uploaded by BakaKaito.
Slightly reformat and compactify catch embeds.
Add missing FixOT DM notification.
Disallow queue entry if item isn't recognized and memes are enabled.
Make backups before writing to Json.
Test TradeCord's Json integrity before making further writes or backups.
DenBot and AutoRoll fixes.
Reluctantly add injection for DenBot as it probably doesn't matter at this point in the games' life cycle.
Add memes as an option for ItemTrade.
Add reactions to sort list in alphabetical order (initial commit, will need changes).
Allow filters for all "list" parameters.
Minor DenBot fixes.
Miscellaneous minor TradeCord changes and fixes.
Add Discord AsyncCommands support to TradeCord (#27)

* Async TradeCord test.

* Gifting fix?

* Fix Json read.
More checks for serialization just in case.

* Gift fix??

* Sort of a gift fix?
Add consideration for Obstagoon in daycare.

* Sort of an async fix.
Change $tis to set info via file upload.
Other changes and fixes.

* Minor gifting and file dumping tweaks.

* More gifting and command queue adjustments.
FixOT hacked mon dumping fix.
Some set generation changes.

* Register gifts to dex if dex hasn't been completed yet.
Add consideration for Slowbro-Galar, Slowking-Galar, Sirfetch'd in daycare.
Add PikaClone event.

* Set the lowest possible level for catches.

* Minor DenBot fix and update.
Met level cannot be higher than current.
Handle wc held items.
Apply Everstone for event trade evos as well.
Band aid fixes for Nidoran in daycare, GO Mew, certain low-level raid encounters.
Add "legendaries" list search for TradeCord.
Fixes for FixOT.
Disallow queue entry if an item for ItemTrade was entered incorrectly.
- Split and change ItemMule and Ditto trades into their own commands and a common role.
- Remove or simplify some boolean checks.
- Extract and move the method for silly memes.
Fix random abilities for older generations and IV randomizer for GO?
Minor embed shiny type fix.
Make LairBot selectable, fix USB-Botbase compatibility via a cheap hack.
Re-add missing LairModule.cs for remote LairBot control via Discord.

+++ SysBot.Pokemon/Util/SeedSearchUtil.cs
@@ -157,12 +157,12 @@ namespace SysBot.Pokemon
                     if (!charMatch)
                         continue;
 
-                    rng = GetAbility(rng, raidInfo.Den.IsEvent ? raidInfo.RaidDistributionEncounter.Ability : (uint)raidInfo.RaidEncounter.Ability, out uint abilityT);
+                    rng = GetAbility(rng, raidInfo.Den.IsEvent ? (uint)raidInfo.RaidDistributionEncounter.Ability : (uint)raidInfo.RaidEncounter.Ability, out uint abilityT);
                     bool abilityMatch = raidInfo.Ability == AbilityType.Any ? abilityT != (uint)raidInfo.Ability : abilityT == (uint)raidInfo.Ability;
                     if (!abilityMatch)
                         continue;
 
-                    rng = GetGender(rng, raidInfo.Ratio, raidInfo.Den.IsEvent ? raidInfo.RaidDistributionEncounter.Gender : (uint)raidInfo.RaidEncounter.Gender, out uint genderT);
+                    rng = GetGender(rng, raidInfo.Ratio, raidInfo.Den.IsEvent ? (uint)raidInfo.RaidDistributionEncounter.Gender : (uint)raidInfo.RaidEncounter.Gender, out uint genderT);
                     bool genderMatch = raidInfo.Gender == GenderType.Any ? genderT != (uint)raidInfo.Gender : genderT == (uint)raidInfo.Gender;
                     if (!genderMatch)
                         continue;
